Creando el modelo EntradasHuacalesDetalle
