+++ P1_AP1_AdonisMercado/Models/EntradasHuacalesDetalle.cs
@@ -1,0 +1,19 @@
+﻿using System.ComponentModel.DataAnnotations;
+using System.ComponentModel.DataAnnotations.Schema;
+
+namespace P1_AP1_AdonisMercado.Models;
+
+public class EntradasHuacalesDetalle
+{
+    [Key]
+    public int DetalleId { get; set; }
+    public int IdEntrada { get; set; }
+    public int TipoId { get; set; }
+    public int Cantidad { get; set; }
+
+    public int Precio { get; set; }
+
+    [ForeignKey("IdEntrada")]
+    [InverseProperty("EntradasHuacalesDetalle")]
+    public virtual EntradasHuacales EntradaHuacal { get; set; }
+}
